Introduce simple MVC structure
