+++ app/core/Controller.php
@@ -1,0 +1,8 @@
+<?php
+class Controller {
+    protected function render(string $view, array $data = []): void
+    {
+        extract($data);
+        require __DIR__ . '/../views/' . $view . '.php';
+    }
+}
