+++ app/models/Bootstrap.php
@@ -1,0 +1,2 @@
+<?php
+require_once __DIR__ . '/../../db_verbindung.php';
